+++ src/pages/helpers/HelperTaskForm.tsx
@@ -41,6 +41,7 @@ import {
   HelperTaskUpdateRequest,
 } from "@/model/helpers-dtos";
 import client from "@/utils/client";
+import { getNow } from "@/utils/date-utils";
 import dayjs from "@/utils/dayjs";
 
 import { canEditTask, getTaskLocation, isMultiDayShift } from "./helpers-utils";
@@ -223,10 +224,54 @@ const HelperTaskForm = ({
       );
     }
 
+    const now = getNow();
+    // Using the first meaningful date is good enough for the confirmation dialog
+    const taskDate = [base.startsAt, base.endsAt, base.deadline].find(
+      (value) => value !== null,
+    );
+    const taskYear = taskDate?.year();
+
+    if (taskYear && taskYear !== now.year()) {
+      confirmations.push(
+        <>
+          Are you sure that{" "}
+          <strong>this task is NOT in the current year</strong>
+          {"?"}
+        </>,
+      );
+    }
+
+    if (taskDate?.isBefore(now)) {
+      confirmations.push(
+        <>
+          Are you sure that <strong>this task is in the past</strong>
+          {"?"} Members cannot sign up for tasks in the past.
+        </>,
+      );
+    }
+
+    if (task?.validatedBy) {
+      confirmations.push(
+        <>
+          Are you sure that you want to{" "}
+          <strong>modify this validated task</strong>
+          {"?"}
+        </>,
+      );
+    } else if (task?.markedAsDoneBy) {
+      confirmations.push(
+        <>
+          Are you sure you want to modify{" "}
+          <strong>this task that has been marked as done</strong>
+          {"?"}
+        </>,
+      );
+    }
+
     if (!base.published) {
       confirmations.push(
         <>
-          Are you sure you want this task to be <strong>unpublished</strong>?
+          Are you sure you want <strong>this task to be unpublished</strong>?
           Members do not see and cannot sign up for unpublished tasks.
         </>,
       );
@@ -319,6 +364,7 @@ const HelperTaskForm = ({
       </SpacedBox>
 
       <SpacedBox>
+        z
         <SpacedTypography fontWeight="bold">
           Long description (not included in emails):
         </SpacedTypography>